brief_gaetan/Création d’une application d’analyse de données sportives via SQL et interface graphique/streamlit/requirements.txt
streamlit
sqlalchemy
sqlalchemy_utils
psycopg2-binary
